requirements.txt
openai-whisper
moviepy>=1.0.3
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
click>=8.1.0
pedalboard